chucky-persona-bot/Chucky AI/requirements.txt
google-genai
python-dotenv
anthropic
elevenlabs
fal-client
requests
openai
mutagen
Pillow
streamlit
